Revert to u64 from u128

+++ src/contractionpath/paths/branchbound.rs
@@ -206,7 +206,7 @@ impl<'a> OptimizePath for BranchBound<'a> {
             }
             self.size_cache
                 .entry(index)
-                .or_insert_with(|| tensor.shape().iter().product::<u128>() as f64);
+                .or_insert_with(|| tensor.shape().iter().product::<u64>() as f64);
 
             self.tensor_cache.entry(index).or_insert_with(|| tensor);
         }
